% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{BIAS}
\alias{BIAS}
\title{Value Bias}
\usage{
BIAS(df)
}
\arguments{
\item{df}{Data.frame with two columns for observation and
modeled value. Additional columns beyond index 1 and 2 will be
ignored.}
}
\description{
Mean overall bias in supplied values. NA values will be removed
}
