% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_drag.R
\name{coef_drag}
\alias{coef_drag}
\title{Coefficient of wind drag}
\usage{
coef_drag(wstr = NULL)
}
\arguments{
\item{wstr}{The wind sheltering coefficient (see \link{wind_sheltering})}

\item{site_id}{The character ID for the requested data}
}
\value{
Coefficient of wind drag
}
\description{
Get coefficient of wind drag for a lake with a supplied wind sheltering coefficient
}
\details{
TODO
}
