% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conical_bathy.R
\name{conical_bathy}
\alias{conical_bathy}
\title{Estimate conical bathymetry}
\usage{
conical_bathy(area, zmax)
}
\arguments{
\item{area}{Lake area in desired units (usually meters^2)}

\item{zmax}{Maximum lake depth (usually in meters)}
}
\description{
Creates simple conical bathymetry profile based
on lake area and max observed depth. Useful
in situations where true hypsometry or
bathymetry is unknown.
}
