% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_drivers.R
\name{driver_aggregate}
\alias{driver_aggregate}
\title{Daily aggregate GLM drivers}
\usage{
driver_aggregate(glm_drivers, avg.window = 1)
}
\arguments{
\item{glm_drivers}{data.frame of GLM formatted drivers}

\item{avg.window}{Rolling window to apply first, in days.}
}

