% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_2_baro.R
\name{elevation_2_baro}
\alias{elevation_2_baro}
\title{Estimate barometric pressure from elevation}
\usage{
elevation_2_baro(elevation)
}
\arguments{
\item{elevation}{Elevation in meters above sea level (MASL)}
}
\value{
Barometric pressure in millibars.
}
\description{
Calculate barometric pressure from elevation. Uses
approximat relationship based on XXX. See references.
}
