% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lw_down.R
\name{est_lw_down}
\alias{est_lw_down}
\title{Estimate downwelling longwave radiation}
\usage{
est_lw_down(dateTime, sw, lat, atm.press, airT, RH)
}
\description{
Estimate downwelling longwave radiation based on atmospheric conditions.
}
