% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{MAE}
\alias{MAE}
\title{Mean average error}
\usage{
MAE(df)
}
\arguments{
\item{df}{Data.frame with two columns for observation and
modeled value. Additional columns beyond index 1 and 2 will be
ignored.}
}
\description{
Mean of the absolute value error for all supplied values.
NA values will be removed
}
