% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsat.R
\name{qsat}
\alias{qsat}
\title{Saturation Specific Humidity}
\usage{
qsat(Ta, Pa)
}
\arguments{
\item{Ta}{air temperature [deg C]}

\item{Pa}{pressure [mb]}
}
\description{
Returns saturation specific humidity [kg/kg] given
 the air temperature and pressure. Useful in convering
 specific humitity to relative humidity
}
