% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{RMSE}
\alias{RMSE}
\title{Root Mean Squared error}
\usage{
RMSE(df)
}
\arguments{
\item{df}{Data.frame with two columns for observation and
modeled value. Additional columns beyond index 1 and 2 will be
ignored.}
}
\description{
Root Mean Squared error of supplied values. NA values will be removed
}

