% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_glm_nml.R
\name{run_glm_nml}
\alias{run_glm_nml}
\title{Directly Run GLM NML file}
\usage{
run_glm_nml(nml, drivers, dest)
}
\arguments{
\item{nml}{An NML object based on object format
used in \code{glmtools}. Required.}

\item{drivers}{A \code{data.frame} of input drivers. To
be saved adjacent to model run (useful for fast temp drives).}

\item{dest}{Temporary destination for model run. Defaults
to a random subfolder of the R \code{tempdir} directory.}
}
\description{
Accepts NML object (and other optional paramters) and
quickly sets up a temporary environment and runs the GLM model.
Returns data.frame row with paths to relevant model files.
}
