% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_sheltering.R
\name{wind_sheltering}
\alias{wind_sheltering}
\title{Calculate wind sheltering coefficient}
\usage{
wind_sheltering(area, canopy = NA, method = "Hondzo")
}
\arguments{
\item{area}{Lake area in meters^2 (square meters)}

\item{canopy}{Surrounding canopy height in meters (for 'Markfort' method)}

\item{method}{The desired calculation method one of c('Markfort', 'Hondzo')}
}
\value{
The wind sheltering coefficient (between 0 and 1)
}
\description{
Calculate the wind sheltering coefficient of a lake.
}
\details{

}
\references{
CD Markfort, Perez AL, Thill JW, Jaster DA, Porté-Agel F, Stefan HG,
2010. Wind sheltering of a lake by a tree canopy or bluff topography. Water Resour. Res. 46.

M Hondzo, Stefan HG, 1993. Lake water temperature simulation model. J. Hydraulic Eng. 119, 1251–1273
}
